Update API interactions
