+++ src/main/java/net/dv8tion/jda/api/interactions/commands/MessageCommandInteraction.java
@@ -1,0 +1,16 @@
+package net.dv8tion.jda.api.interactions.commands;
+
+import net.dv8tion.jda.api.entities.Message;
+
+import javax.annotation.Nonnull;
+
+public interface MessageCommandInteraction extends CommandInteraction
+{
+    /**
+     * The message clicked on.
+     *
+     * @return The message clicked on
+     */
+    @Nonnull
+    Message getInteractedMessage();
+}
